requirements.txt
discord.py>=2.4.0,<3.0.0
python-whois>=0.9.4,<1.0.0
